feat: integrated prototype repeatRenderModule
